changing header
changing header and trying to make a sticky header function.

The header looks neater now, but I  got to do more work on the sticky header function for it to  work.

I also have to make a function that closes the header if I press an item  inside of the menu  (on mobile)